feat: 체스 게임을 진행하는 클래스 구현
